requirements.txt
Flask
python-dotenv
requests
paho-mqtt
pytest
pytest-mock
coverage
PyJWT
flasgger
